[backend] v4
Squashed:
[backend] Fix node engine version in package.json

[frontend] More lost changes on rebase

[frontend] Lost changes on rebase

Add mongo at build time (for agenda)

[frontend] lint

[*] Fix leith's code ideology

[backend] Fix leith's mess

[backend] Fix broken tests

[backend] Better

[backend] Add id checks

+++ server/services/default-meta.mjs
@@ -29,14 +29,20 @@ export default function(Model) {
             return await query.exec();
         }
         static async get(id, proj, population) {
+            if(!Types.ObjectId.isValid(id))
+                return null;
             const query = Model.findById(id, proj);
             populate(query, population);
             return await query.exec();
         }
         static async update(id, content) {
+            if(!Types.ObjectId.isValid(id))
+                return null;
             return await Model.findByIdAndUpdate(id, content, {'new': true}).exec();
         }
         static async remove(id) {
+            if(!Types.ObjectId.isValid(id))
+                return null;
             return await Model.findByIdAndRemove(id).exec();
         }
         static async prune(cond) {
